Move FileBaton to its own file
